requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
deepdiff==8.2.0
idna==3.10
mypy==1.15.0
mypy-extensions==1.0.0
numpy==2.2.3
orderly-set==5.3.0
pandas-stubs==2.2.3.241126
PyYAML==6.0.2
requests==2.32.3
types-pytz==2025.1.0.20250204
types-requests==2.32.0.20250301
typing_extensions==4.12.2
urllib3==2.3.0
